+++ tests/Enums/MonthEnum.php
@@ -2,18 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Spatie\Enum\Tests\TestClasses;
+namespace Spatie\Enum\Tests\Enums;
 
 use Spatie\Enum\Enum;
 
 abstract class MonthEnum extends Enum
 {
-    abstract public function getNumericIndex(): int;
-
     public static function january(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 1;
             }
@@ -23,7 +21,7 @@ abstract class MonthEnum extends Enum
     public static function february(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 2;
             }
@@ -33,7 +31,7 @@ abstract class MonthEnum extends Enum
     public static function march(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 3;
             }
@@ -43,7 +41,7 @@ abstract class MonthEnum extends Enum
     public static function april(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 4;
             }
@@ -53,7 +51,7 @@ abstract class MonthEnum extends Enum
     public static function may(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 5;
             }
@@ -63,7 +61,7 @@ abstract class MonthEnum extends Enum
     public static function june(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 6;
             }
@@ -73,7 +71,7 @@ abstract class MonthEnum extends Enum
     public static function july(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 7;
             }
@@ -83,7 +81,7 @@ abstract class MonthEnum extends Enum
     public static function august(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 8;
             }
@@ -93,7 +91,7 @@ abstract class MonthEnum extends Enum
     public static function september(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 9;
             }
@@ -103,7 +101,7 @@ abstract class MonthEnum extends Enum
     public static function october(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 10;
             }
@@ -113,7 +111,7 @@ abstract class MonthEnum extends Enum
     public static function november(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 11;
             }
@@ -123,7 +121,7 @@ abstract class MonthEnum extends Enum
     public static function december(): MonthEnum
     {
         return new class() extends MonthEnum {
-            public function getNumericIndex(): int
+            public function getIndex(): int
             {
                 return 12;
             }